refactor(ui): split InfrastructureModal.tsx god object into modules
