+++ scripts/.god-object-baseline.txt
@@ -11,7 +11,6 @@
 738 ui/src/routes/sandboxes.$id.tsx
 721 ui/src/components/shared/SessionSidebar.tsx
 701 ui/src/lib/stores/chatSessions.ts
-682 ui/src/components/shared/InfrastructureModal.tsx
 653 ai-agent-instance-blueprint-lib/src/auto_provision.rs
 627 ui/src/lib/hooks/useCreateDeploy.ts
 624 ui/src/lib/api/sandboxClient.ts
